Disable gyroscope when screen not in focus
Related to #1252
