+++ apps/mobile-wallet/src/components/animatedBackground/animatedBackgroundTypes.ts
@@ -1,0 +1,12 @@
+import { SharedValue } from 'react-native-reanimated'
+
+export interface AnimatedBackgroundProps {
+  height?: number
+  width?: number
+  scrollY?: SharedValue<number>
+  isAnimated?: boolean
+  usesGyroscope?: boolean
+  isFullScreen?: boolean
+  showAlephiumLogo?: boolean
+  shade?: string
+}
